feat: refactor listeners and add config
BREAKING CHANGE: added config to ws client
