+++ src/worldstate/client.rs
@@ -2,19 +2,35 @@
 
 //! A client to do all sorts of things with the API
 
+use std::time::Duration;
+
 use reqwest::StatusCode;
 
 use super::{
     Queryable,
-    TimedEvent,
     error::Error,
     language::Language,
     models::items::Item,
-    utils::{
-        Change,
-        CrossDiff,
-    },
 };
+
+#[derive(Debug, Clone)]
+pub struct ClientConfig {
+    /// The time a nested listener should sleep before making another request
+    pub nested_listener_sleep: Duration,
+
+    /// The time a listener sleeps upon reaching it's expiry until it tries to fetch the updated
+    /// data
+    pub listener_sleep_timeout: Duration,
+}
+
+impl Default for ClientConfig {
+    fn default() -> Self {
+        Self {
+            nested_listener_sleep: Duration::from_mins(2),
+            listener_sleep_timeout: Duration::from_mins(1),
+        }
+    }
+}
 
 /// The client that acts as a convenient way to query models.
 ///
@@ -43,8 +59,9 @@ use super::{
 /// Check the [queryable](crate::worldstate::queryable) module for all queryable types.
 #[derive(Debug, Clone)]
 pub struct Client {
-    http: reqwest::Client,
-    base_url: String,
+    pub(crate) http: reqwest::Client,
+    pub(crate) base_url: String,
+    pub(crate) config: ClientConfig,
 }
 
 impl Default for Client {
@@ -55,6 +72,7 @@ impl Default for Client {
         Self {
             http: reqwest::Client::new(),
             base_url: "https://api.warframestat.us".to_string(),
+            config: ClientConfig::default(),
         }
     }
 }
@@ -62,16 +80,14 @@ impl Default for Client {
 impl Client {
     /// Creates a new [Client] with the option to supply a custom reqwest client and a base url.
     #[must_use]
-    pub fn new(reqwest_client: reqwest::Client, base_url: String) -> Self {
+    pub fn new(reqwest_client: reqwest::Client, base_url: String, config: ClientConfig) -> Self {
         Self {
             http: reqwest_client,
             base_url,
+            config,
         }
     }
-}
 
-// impl FETCH
-impl Client {
     /// Fetches an instance of a specified model.
     ///
     /// # Example
@@ -222,384 +238,5 @@ impl Client {
         let item = serde_json::from_str::<Item>(&json)?;
 
         Ok(Some(item))
-    }
-
-    /// Asynchronous method that continuously fetches updates for a given type `T` and invokes a
-    /// callback function.
-    ///
-    /// # Arguments
-    ///
-    /// - `callback`: A function that implements the `ListenerCallback` trait and is called with the
-    ///   previous and new values of `T`.
-    ///
-    /// # Generic Constraints
-    ///
-    /// - `T`: Must implement the `Queryable` and `TimedEvent` traits.
-    /// - `Callback`: Must implement the `ListenerCallback` trait with a lifetime parameter `'any`
-    ///   and type parameter `T`.
-    ///
-    /// # Returns
-    ///
-    /// - `Result<(), Error>`: Returns `Ok(())` if the operation is successful, otherwise returns an
-    ///   `Error`.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use std::error::Error;
-    ///
-    /// use warframe::worldstate::{
-    ///     Client,
-    ///     queryable::Cetus,
-    /// };
-    ///
-    /// async fn on_cetus_update(before: &Cetus, after: &Cetus) {
-    ///     println!("BEFORE : {before:?}");
-    ///     println!("AFTER  : {after:?}");
-    /// }
-    ///
-    /// #[tokio::main]
-    /// async fn main() -> Result<(), Box<dyn Error>> {
-    ///     let client = Client::default();
-    ///     
-    ///     client.call_on_update(on_cetus_update); // don't forget to start it as a bg task (or .await it)s
-    ///     Ok(())
-    /// }
-    /// ```
-    #[allow(clippy::missing_panics_doc)]
-    pub async fn call_on_update<T, Callback>(&self, callback: Callback) -> Result<(), Error>
-    where
-        T: TimedEvent + Queryable<Return = T>,
-        for<'a, 'b> Callback: AsyncFn(&'a T, &'b T),
-    {
-        tracing::debug!("{} (LISTENER) :: Started", std::any::type_name::<T>());
-        let mut item = self.fetch::<T>().await?;
-
-        loop {
-            if item.expiry() <= chrono::offset::Utc::now() {
-                tracing::debug!(
-                    listener = %std::any::type_name::<T>(),
-                    "(LISTENER) Fetching new possible update"
-                );
-
-                tokio::time::sleep(std::time::Duration::from_secs(30)).await;
-
-                let new_item = self.fetch::<T>().await?;
-
-                if item.expiry() >= new_item.expiry() {
-                    continue;
-                }
-
-                callback(&item, &new_item).await;
-                item = new_item;
-            }
-
-            let time_to_sleep = item.expiry() - chrono::offset::Utc::now();
-
-            tracing::debug!(
-                listener = %std::any::type_name::<T>(),
-                sleep_duration = %time_to_sleep.num_seconds(),
-                "(LISTENER) Sleeping"
-            );
-
-            tokio::time::sleep(time_to_sleep.to_std().unwrap()).await;
-        }
-    }
-
-    /// Asynchronous method that continuously fetches updates for a given type `T` and invokes a
-    /// callback function.
-    ///
-    /// # Arguments
-    ///
-    /// - `callback`: A function that implements the `ListenerCallback` trait and is called with the
-    ///   previous and new values of `T`.
-    ///
-    /// # Generic Constraints
-    ///
-    /// - `T`: Must implement the `Queryable`, `TimedEvent` and `PartialEq` traits.
-    /// - `Callback`: Must implement the `ListenerCallback` trait with a lifetime parameter `'any`
-    ///   and type parameter `T`.
-    ///
-    /// # Returns
-    ///
-    /// - `Result<(), Error>`: Returns `Ok(())` if the operation is successful, otherwise returns an
-    ///   `Error`.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use std::error::Error;
-    ///
-    /// use warframe::worldstate::{
-    ///     Client,
-    ///     Change,
-    ///     queryable::Fissure,
-    /// };
-    ///
-    /// /// This function will be called once a fissure updates.
-    /// /// This will send a request to the corresponding endpoint once every 30s
-    /// /// and compare the results for changes.
-    /// async fn on_fissure_update(fissure: &Fissure, change: Change) {
-    ///     match change {
-    ///         Change::Added => println!("Fissure ADDED   : {fissure:?}"),
-    ///         Change::Removed => println!("Fissure REMOVED : {fissure:?}"),
-    ///     }
-    /// }
-    ///
-    /// #[tokio::main]
-    /// async fn main() -> Result<(), Box<dyn Error>> {
-    ///     // initialize a client (included in the prelude)
-    ///     let client = Client::default();
-    ///
-    ///     // Pass the function to the handler
-    ///     // (will return a Future)
-    ///     client.call_on_nested_update(on_fissure_update); // don't forget to start it as a bg task (or .await it)
-    ///     Ok(())
-    /// }
-    /// ```
-    #[allow(clippy::missing_panics_doc)]
-    #[allow(clippy::missing_errors_doc)]
-    pub async fn call_on_nested_update<T, Callback>(&self, callback: Callback) -> Result<(), Error>
-    where
-        T: TimedEvent + Queryable<Return = Vec<T>> + PartialEq,
-        for<'any> Callback: AsyncFn(&'any T, Change),
-    {
-        tracing::debug!(
-            listener = %std::any::type_name::<Vec<T>>(),
-            "(LISTENER) Started"
-        );
-
-        let mut items = self.fetch::<T>().await?;
-
-        loop {
-            tokio::time::sleep(std::time::Duration::from_secs(30)).await;
-
-            tracing::debug!(
-                listener = %std::any::type_name::<Vec<T>>(),
-                "(LISTENER) Fetching new possible state"
-            );
-
-            let new_items = self.fetch::<T>().await?;
-
-            let diff = CrossDiff::new(&items, &new_items);
-
-            let removed_items = diff.removed();
-            let added_items = diff.added();
-
-            if !removed_items.is_empty() || !added_items.is_empty() {
-                tracing::debug!(
-                    listener = %std::any::type_name::<Vec<T>>(),
-                    "(LISTENER) Found changes, proceeding to call callback with every change"
-                );
-
-                for (item, change) in removed_items.into_iter().chain(added_items) {
-                    // call callback fn
-                    callback(item, change).await;
-                }
-                items = new_items;
-            }
-        }
-    }
-
-    /// Asynchronous method that calls a callback function with state on update.
-    ///
-    /// # Arguments
-    ///
-    /// - `callback`: A callback function that takes the current item, the new item, and the state
-    ///   as arguments.
-    /// - `state`: The state object that will be passed to the callback function.
-    ///
-    /// # Generic Parameters
-    ///
-    /// - `S`: The type of the state object. It must be `Sized`, `Send`, `Sync`, and `Clone`.
-    /// - `T`: Must implement the `Queryable` and `TimedEvent` traits.
-    /// - `Callback`: The type of the callback function. It must implement the
-    ///   `StatefulListenerCallback` trait with the item type `T` and the state type `S`.
-    ///
-    /// # Returns
-    ///
-    /// This method returns a `Result` indicating whether the operation was successful or an
-    /// `Error` occurred. The result is `Ok(())` if the operation was successful.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use std::{error::Error, sync::Arc};
-    ///
-    /// use warframe::worldstate::{Client, queryable::Cetus};
-    ///
-    /// // Define some state
-    /// #[derive(Debug)]
-    /// struct MyState {
-    ///     _num: i32,
-    ///     _s: String,
-    /// }
-    ///
-    /// async fn on_cetus_update(state: Arc<MyState>, before: &Cetus, after: &Cetus) {
-    ///     println!("STATE  : {state:?}");
-    ///     println!("BEFORE : {before:?}");
-    ///     println!("AFTER  : {after:?}");
-    /// }
-    ///
-    /// #[tokio::main]
-    /// async fn main() -> Result<(), Box<dyn Error>> {
-    ///     let client = Client::default();
-    ///
-    ///     // Note that the state will be cloned into the handler, so Arc is preferred
-    ///     let state = Arc::new(MyState {
-    ///         _num: 69,
-    ///         _s: "My ginormous ass".into(),
-    ///     });
-    ///
-    ///     client
-    ///         .call_on_update_with_state(on_cetus_update, state); // don't forget to start it as a bg task (or .await it)
-    ///     Ok(())
-    /// }
-    /// ```
-    #[allow(clippy::missing_panics_doc)]
-    pub async fn call_on_update_with_state<S, T, Callback>(
-        &self,
-        callback: Callback,
-        state: S,
-    ) -> Result<(), Error>
-    where
-        S: Sized + Send + Sync + Clone,
-        T: TimedEvent + Queryable<Return = T>,
-        for<'a, 'b> Callback: AsyncFn(S, &'a T, &'b T),
-    {
-        let mut item = self.fetch::<T>().await?;
-
-        loop {
-            if item.expiry() <= chrono::offset::Utc::now() {
-                tracing::debug!(
-                    listener = %std::any::type_name::<T>(),
-                    "(LISTENER) Fetching new possible state"
-                );
-
-                tokio::time::sleep(std::time::Duration::from_secs(30)).await;
-
-                let new_item = self.fetch::<T>().await?;
-
-                if item.expiry() >= new_item.expiry() {
-                    continue;
-                }
-                callback(state.clone(), &item, &new_item).await;
-                item = new_item;
-            }
-
-            let time_to_sleep = item.expiry() - chrono::offset::Utc::now();
-
-            tracing::debug!(
-                listener = %std::any::type_name::<T>(),
-                sleep_duration = %time_to_sleep.num_seconds(),
-                "(LISTENER) Sleeping"
-            );
-
-            tokio::time::sleep(time_to_sleep.to_std().unwrap()).await;
-        }
-    }
-
-    /// Asynchronous method that calls a callback function on nested updates with a given state.
-    ///
-    /// # Arguments
-    ///
-    /// * `callback` - The callback function to be called on each change.
-    /// * `state` - The state to be passed to the callback function.
-    ///
-    /// # Generic Constraints
-    ///
-    /// * `S` - The type of the state, which must be `Sized`, `Send`, `Sync`, and `Clone`.
-    /// * `T` - Must implement the `Queryable`, `TimedEvent` and `PartialEq` traits.
-    /// * `Callback` - The type of the callback function, which must implement the
-    ///   `StatefulNestedListenerCallback` trait.
-    ///
-    /// # Returns
-    ///
-    /// Returns `Ok(())` if the callback function is successfully called on each change, or an
-    /// `Error` if an error occurs.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use std::{error::Error, sync::Arc};
-    ///
-    /// use warframe::worldstate::{Change, Client, queryable::Fissure};
-    ///
-    /// // Define some state
-    /// #[derive(Debug)]
-    /// struct MyState {
-    ///     _num: i32,
-    ///     _s: String,
-    /// }
-    ///
-    /// async fn on_fissure_update(state: Arc<MyState>, fissure: &Fissure, change: Change) {
-    ///     println!("STATE  : {state:?}");
-    ///     match change {
-    ///         Change::Added => println!("FISSURE ADDED   : {fissure:?}"),
-    ///         Change::Removed => println!("FISSURE REMOVED : {fissure:?}"),
-    ///     }
-    /// }
-    ///
-    /// #[tokio::main]
-    /// async fn main() -> Result<(), Box<dyn Error>> {
-    ///     let client = Client::default();
-    ///
-    ///     // Note that the state will be cloned into the handler, so Arc is preferred
-    ///     let state = Arc::new(MyState {
-    ///         _num: 69,
-    ///         _s: "My ginormous ass".into(),
-    ///     });
-    ///
-    ///     client
-    ///         .call_on_nested_update_with_state(on_fissure_update, state); // don't forget to start it as a bg task (or .await it)
-    ///     Ok(())
-    /// }
-    /// ```
-    #[allow(clippy::missing_panics_doc)]
-    pub async fn call_on_nested_update_with_state<S, T, Callback>(
-        &self,
-        callback: Callback,
-        state: S,
-    ) -> Result<(), Error>
-    where
-        S: Sized + Send + Sync + Clone,
-        T: Queryable<Return = Vec<T>> + TimedEvent + PartialEq,
-        for<'any> Callback: AsyncFn(S, &'any T, Change),
-    {
-        tracing::debug!(
-            listener = %std::any::type_name::<Vec<T>>(),
-            "(LISTENER) Started"
-        );
-
-        let mut items = self.fetch::<T>().await?;
-
-        loop {
-            tokio::time::sleep(std::time::Duration::from_secs(30)).await;
-
-            tracing::debug!(
-                listener = %std::any::type_name::<Vec<T>>(),
-                "(LISTENER) Fetching new possible state"
-            );
-
-            let new_items = self.fetch::<T>().await?;
-
-            let diff = CrossDiff::new(&items, &new_items);
-
-            let removed_items = diff.removed();
-            let added_items = diff.added();
-
-            if !removed_items.is_empty() || !added_items.is_empty() {
-                tracing::debug!(
-                    listener = %std::any::type_name::<Vec<T>>(),
-                    "(LISTENER) Found changes, proceeding to call callback with every change"
-                );
-
-                for (item, change) in removed_items.into_iter().chain(added_items) {
-                    // call callback fn
-                    callback(state.clone(), item, change).await;
-                }
-                items = new_items;
-            }
-        }
     }
 }
