+++ src/worldstate/mod.rs
@@ -30,6 +30,8 @@ mod models;
 pub mod utils;
 
 pub mod language;
+mod listener;
+mod listener_nested;
 
 /// A module that re-exports every type that is queryable
 pub mod queryable {
